Added instances file for all classes other than the player Started working on shop and it's design Cleaned some code
