+++ Main.py
@@ -1,9 +1,11 @@
 import time, sys, os, random, json
-from rich import print as rprint
 import rich , Engine
 from datetime import date
 from Engine import Game
 
+from Instances import *
+
+#----Window-Setup-----#
 Game.disable_quickedit()
 Game.window('Game',50,40)
 
@@ -12,21 +14,22 @@ Engine.console = Engine.Console()
 
 Engine.layout.split_column(
     Engine.Layout(name='Main',ratio=1),
-    Engine.Layout(name='Side',ratio=2)
+    Engine.Layout(name='SideSplit',ratio=2)
 )
-
+Engine.layout["SideSplit"].split_row(
+    Engine.Layout(name='Side',ratio=1),
+    Engine.Layout(name='Shop',ratio=1,visible=False)
+)
+#----Window-Setup-----#
 
 #---------
-GameVersion = "0.0.6"
+GameVersion = "0.0.7"
 SaveFilePath = os.getenv('APPDATA')+"\SuperNiceGame"
-try: 
-    os.mkdir(SaveFilePath) 
-except OSError as error: 
-    pass
+if not os.path.exists(SaveFilePath):
+    os.mkdir(SaveFilePath)
 HIGH_LIGHT_COLOR = "green"
 PANEL_COLOR = "bright_white"
 #--------- # TODO ADD OPTION MENU WITH OPTION TO CHANGE COLORS AND ENABLE OR DISABLE AUTOSAVE AND MORE IN THE FUTURE
-
 
 class Player:
     def __init__(self):
@@ -58,7 +61,7 @@ class Player:
         if len(player.Inventory) <= 0: return
         for x in player.Inventory:
             if Name == x.name:
-                x.Use()
+                x.Use(self)
                 break
         self.RecalculateStatsFromEquipment()
     
@@ -89,6 +92,7 @@ class Player:
                     Engine.layout['Side'].update(Engine.Panel(SideText,title="Inventory",style="%s" %PANEL_COLOR))
                     Game.wait_for_input()
                     return
+                if self.GetHealth() >= self.GetMaxHealth(): return
                 self.Potions[Name] -= 1
                 HealthBefore = round(player.GetHealth())
                 Heal = round(player.GetMaxHealth() * self.PotionHealProcent)
@@ -101,7 +105,7 @@ class Player:
                 Engine.layout['Side'].update(Engine.Panel(SideText,title="Inventory",style="%s" %PANEL_COLOR))
                 Game.wait_for_input()
     
-    def SetAllStats(self,NewStats,NewClass): #TODO load class
+    def SetAllStats(self,NewStats,NewClass):
         self.Class = NewClass
         self.Stats = NewStats
     
@@ -180,598 +184,8 @@ class Player:
     def SetDefence(self,Value):
         self.Stats['Defence'] = Value
 
-class CharacterClass:
-
-    Classes = []
-
-    def __init__(self,Name:str,MaxHealth:int,Attack:int,Defence:int):
-        self.Name = Name
-        self.Stats = {
-            'Health' : MaxHealth,
-            'MaxHealth' : MaxHealth,
-            'Attack' : Attack,
-            'Defence' : Defence
-        }
-        CharacterClass.Classes.append(self)
-
-class LootTable:
-    def __init__(self):
-        self.Loot = []
-        self.PossibleLoot = []
-    
-    def addLoot(self,Item,Weight:int):
-        self.Drop = {
-            'Item': Item,
-            'Weight' : Weight
-        }
-        self.PossibleLoot.append(self.Drop)
-
-    def Roll(self):
-        self.Loot = []
-        self.AllWeights = 0
-        self.Dropped = None
-        self.previous = None
-        for x in self.PossibleLoot:
-            self.LootItem = {
-                'Item': None,
-                'MinChance': 0,
-                'MaxChance': 0
-            }
-            self.AllWeights += x['Weight']
-            self.LootItem['Item'] = x['Item']
-            if self.previous == None:
-                self.LootItem['MinChance'] = 1
-                self.LootItem['MaxChance'] = x['Weight']
-            else:
-                self.LootItem['MinChance'] = (self.previous['MaxChance'] + 1)
-                self.LootItem['MaxChance'] = (self.previous['MaxChance'] + x['Weight'])
-            self.Loot.append(self.LootItem)
-            self.previous = self.LootItem
-        rng = random.randint(1,self.AllWeights)
-        for x in self.Loot:
-            if rng <= x['MaxChance'] and rng >= x['MinChance']:
-                if x['Item'] != None:
-                    self.Dropped = x
-        if self.Dropped != None:
-            player.Inventory.append(self.Dropped['Item'])
-        if self.Dropped == None:
-            return None
-        return self.Dropped['Item'].name
-
-class Item:
-
-    ItemBase = []
-
-    def __init__(self,Name:str,Price:int,Damage:int,Defence:int,MaxHealth:int,EquipPlace:str,ClassUse=[]):
-        self.name = Name
-        self.Price = Price
-        self.Damage = Damage
-        self.Defence = Defence
-        self.MaxHealth = MaxHealth
-        self.EquipPlace = EquipPlace
-        # if 'ClassUse' is empty that means every class can use that item
-        self.ClassUse = ClassUse
-        Item.ItemBase.append(self)
-    
-    def Use(self):
-        SideText = Engine.Text("")
-        if len(self.ClassUse) != 0:
-            ClassesToCheck = len(self.ClassUse)
-            for Class in self.ClassUse:
-                if Class.Name == player.Class:
-                    break
-                ClassesToCheck -= 1
-                if ClassesToCheck == 0:
-                    SideText.append("%s is not for your class : %s" % (self.name, player.Class))
-                    Engine.layout['Side'].update(Engine.Panel(SideText,title="Inventory",style="%s" %PANEL_COLOR))
-                    time.sleep(1)
-                    return
-        if player.Equiped[self.EquipPlace] == None:
-            player.Equiped[self.EquipPlace] = self.name
-            player.Inventory.remove(self)
-            SideText.append("You've used %s" % self.name)
-        else:
-            SideText.append("You've exchanged a %s for %s" % (player.Equiped[self.EquipPlace], self.name))
-            player.Inventory.append(Item.FindItem(player.Equiped[self.EquipPlace]))
-            player.Equiped[self.EquipPlace] = self.name
-            player.Inventory.remove(self)
-        player.ResetStatsFromEquipment()
-        player.AddStatsFromEquipment()
-        Engine.layout['Side'].update(Engine.Panel(SideText,title="Inventory",style="%s" %PANEL_COLOR))
-        time.sleep(1)
-    
-    def Sell(self):
-        if len(player.Inventory) <= 0: return
-        for index , _ in enumerate(player.Inventory):
-            if player.Inventory[index] == self:
-                player.Gold += self.Price
-                player.Inventory.remove(self)
-
-    def FindItem(ItemName):
-        for x in Item.ItemBase:
-            if x.name == ItemName:
-                return x
-        return None
-
-class Monster:
-
-    MonsterBase = []
-
-    def __init__(self,Name:str,MaxHealth:int,Attack:int,Defence:int,LootTable:LootTable):
-        self.name = Name
-        self.Stats = {
-            'Health' : MaxHealth,
-            'MaxHealth' : MaxHealth,
-            'Attack' : Attack,
-            'Defence' : Defence
-        }
-        self.LootTable = LootTable
-        Monster.MonsterBase.append(self)
-    
-    def GetHealth(self):
-        return self.Stats['Health']
-    
-    def SetHealth(self,value):
-        self.Stats['Health'] = value
-        if self.Stats['Health'] > self.Stats['MaxHealth']: self.Stats['Health'] = self.Stats['MaxHealth']
-        if self.Stats['Health'] < 0: self.Stats['Health'] = 0
-    
-    def SetMaxHealth(self):
-        self.Stats['Health'] = self.Stats['MaxHealth']
-
-
-#=========Instances=========
 player = Player()
-#----Classes---
-WarriorClass = CharacterClass('Warrior',120,30,20)
-RangerClass = CharacterClass('Ranger',100,25,10)
-MageClass = CharacterClass('Mage',70,40,5)
-TitanClass = CharacterClass('Titan',1000,1000,1000)
-#----Items-----
 
-Helmet = 'Helmet'
-Chestplate = 'Chestplate'
-Leggins = 'Leggins'
-LeftHand = 'LeftHand'
-RightHand = 'RightHand'
-
-#---Leather---
-LeatherHelmet = Item('Leather Helmet',10,0,5,0,Helmet)
-LeatherChestplate = Item('Leather Chestplate',10,0,7,5,Chestplate)
-LeatherLeggins = Item('Leather Leggins',10,0,5,0,Leggins)
-LeatherGloves = Item('Leather Gloves',10,0,5,0,LeftHand)
-#---Iron---
-IronHelmet = Item('Iron Helmet',10,0,5,5,Helmet,[WarriorClass])
-IronChestplate = Item('Iron Chestplate',10,0,30,10,Chestplate,[WarriorClass])
-IronLeggins = Item('Iron Leggins',10,0,10,0,Leggins,[WarriorClass])
-IronShield = Item('Iron Shield',10,-10,20,0,LeftHand,[WarriorClass])
-IronSword = Item('Iron Sword',10,10,0,0,RightHand,[WarriorClass])
-IronDagger = Item('Iron Dagger',10,7,0,0,RightHand,[RangerClass,MageClass])
-#---Wooden---
-WoodenShield = Item('Wooden Shield',10,-5,10,0,LeftHand,[WarriorClass])
-WoodenSword = Item('Wooden Sword',10,5,0,0,RightHand,[WarriorClass])
-WoodenBow = Item('Wooden Bow',10,10,0,0,RightHand,[RangerClass])
-#---Magic---
-BegginersWand = Item('Begginers Wand',10,10,0,0,RightHand,[MageClass])
-StudentCape = Item('Student Cape',10,5,5,5,Chestplate,[MageClass])
-WizardHat = Item('Wizard Hat',10,5,5,0,Helmet,[MageClass])
-MagicTome = Item('Magic Tome',10,10,0,0,LeftHand,[MageClass])
-MagicOrb = Item('Magic Orb',10,20,0,0,LeftHand,[MageClass])
-#---Other---
-Chainmail = Item('Chainmail',10,0,10,5,Chestplate,[RangerClass,WarriorClass])
-
-#----Zombie----
-ZombieLoot = LootTable()
-ZombieLoot.addLoot(WoodenSword,10)
-ZombieLoot.addLoot(WoodenShield,9)
-ZombieLoot.addLoot(IronSword,4)
-ZombieLoot.addLoot(IronDagger,7)
-ZombieLoot.addLoot(IronShield,4)
-ZombieLoot.addLoot(Chainmail,8)
-ZombieLoot.addLoot(None,100)
-Zombie = Monster('Zombie',60,15,20,ZombieLoot)
-#----Warewolf----
-WarewolfLoot = LootTable()
-WarewolfLoot.addLoot(IronHelmet,7)
-WarewolfLoot.addLoot(IronChestplate,5)
-WarewolfLoot.addLoot(IronLeggins,7)
-WarewolfLoot.addLoot(WoodenBow,10)
-WarewolfLoot.addLoot(BegginersWand,10)
-WarewolfLoot.addLoot(IronDagger,10)
-WarewolfLoot.addLoot(IronSword,7)
-WarewolfLoot.addLoot(None,100)
-Warewolf = Monster("Warewolf",50,20,20,WarewolfLoot)
-#----Skeleton----
-SkeletonLoot = LootTable()
-SkeletonLoot.addLoot(Chainmail,10)
-SkeletonLoot.addLoot(LeatherHelmet,9)
-SkeletonLoot.addLoot(LeatherLeggins,5)
-SkeletonLoot.addLoot(LeatherGloves,7)
-SkeletonLoot.addLoot(LeatherChestplate,4)
-SkeletonLoot.addLoot(WoodenBow,8)
-SkeletonLoot.addLoot(None,100)
-Skeleton = Monster('Skeleton',45,10,10,SkeletonLoot)
-#-----Ghost-----
-GhostLoot = LootTable()
-GhostLoot.addLoot(BegginersWand,5)
-GhostLoot.addLoot(WizardHat,9)
-GhostLoot.addLoot(StudentCape,10)
-GhostLoot.addLoot(MagicTome,8)
-GhostLoot.addLoot(None,100)
-Ghost = Monster('Ghost',15,5,500,GhostLoot)
-#-----Thief-----
-ThiefLoot = LootTable()
-ThiefLoot.addLoot(LeatherHelmet,6)
-ThiefLoot.addLoot(LeatherChestplate,4)
-ThiefLoot.addLoot(LeatherLeggins,5)
-ThiefLoot.addLoot(LeatherGloves,7)
-ThiefLoot.addLoot(IronDagger,3)
-ThiefLoot.addLoot(None,100)
-Thief = Monster('Thief',25,10,5,ThiefLoot)
-#-----Wraith-----
-WraithLoot = LootTable()
-WraithLoot.addLoot(BegginersWand,9)
-WraithLoot.addLoot(StudentCape,7)
-WraithLoot.addLoot(MagicTome,8)
-WraithLoot.addLoot(MagicOrb,5)
-WraithLoot.addLoot(None,100)
-Wraith = Monster('Wraith',100,15,5,WraithLoot)
-#-----Wraith-----
-#===========================
-
-def ShowInventory():    # TODO ADD SHOWING ITEM'S STATS ON A MAIN SCREEN WHILE IT IS TARGETED
-    TargetOption = 0
-    TargetItem = None
-    while True:
-        ExitIndex = len(player.Potions) + len(player.Inventory)
-        SideText = Engine.Text("")
-        SideText.append("Potions: \n")
-        for index, name in enumerate(player.Potions):
-            if index == TargetOption: 
-                SideText.append("> %s - %s\n" % (name,player.Potions[name]) ,style="u %s" %HIGH_LIGHT_COLOR)
-                TargetItem = name
-            else: SideText.append("%s - %s\n" % (name,player.Potions[name]))
-        SideText.append("Items: \n")
-        if len(player.Inventory) > 0:
-            for index, Item in enumerate(player.Inventory):
-                if index+len(player.Potions) == TargetOption: 
-                    SideText.append("> %s \n" % Item.name ,style="u %s" %HIGH_LIGHT_COLOR)
-                    TargetItem = Item.name
-                else: SideText.append("%s \n" % Item.name)
-            
-        if TargetOption == ExitIndex: SideText.append("> Go back \n" ,style="u %s" %HIGH_LIGHT_COLOR)
-        else: SideText.append("Go back \n")
-
-        Engine.layout['Main'].update(Engine.Panel(player.PrintStats(),title="Player",style="%s" %PANEL_COLOR))
-        Engine.layout['Side'].update(Engine.Panel(SideText,title="Inventory",style="%s" %PANEL_COLOR))
-        match Game.get_input():
-            case 'w' :
-                TargetOption -= 1
-                if TargetOption < 0: TargetOption = ExitIndex
-            case 's' :
-                TargetOption += 1
-                if TargetOption > ExitIndex: TargetOption = 0
-            case '\r' | ' ':
-                if TargetOption == ExitIndex: return
-                if TargetItem != None:
-                    player.UsePotion(TargetItem)
-                    player.UseItem(TargetItem)
-
-                
-
-def ShowEquipment():
-    TargetOption = 0
-    TargetItem = None
-    ExitIndex = len(player.Equiped)
-    while True:
-        EquipmentText = Engine.Text("")
-        for index , name in enumerate(player.Equiped):
-            Item = player.Equiped[name]
-            if TargetOption == index:
-                EquipmentText.append("> %s - %s \n" %(name,Item) ,style="u %s" %HIGH_LIGHT_COLOR)
-                TargetItem = Item
-            else: EquipmentText.append("%s - %s \n" %(name,Item))
-        if TargetOption == ExitIndex: EquipmentText.append("> Go back" ,style="u %s" %HIGH_LIGHT_COLOR)
-        else: EquipmentText.append("Go back")
-        Engine.layout['Main'].update(Engine.Panel(player.PrintStats(),title="Player",style="%s" %PANEL_COLOR))
-        Engine.layout['Side'].update(Engine.Panel(EquipmentText,title="Equipement",style="%s" %PANEL_COLOR))
-        match Game.get_input():
-            case 'w' :
-                TargetOption -= 1
-                if TargetOption < 0: TargetOption = ExitIndex
-            case 's' :
-                TargetOption += 1
-                if TargetOption > ExitIndex: TargetOption = 0
-            case '\r' | ' ':
-                if TargetOption == ExitIndex: return
-                if TargetItem != None:
-                    Engine.layout['Side'].update(Engine.Panel(player.RemoveEquipment(TargetItem),title="Equipement",style="%s" %PANEL_COLOR))
-                    time.sleep(1)
-
-def ChooseCharacter():
-    Engine.layout['Side'].update(Engine.Panel(Engine.Text("Choose your character!"),style="%s" %PANEL_COLOR))
-    TargetCharacter = 0
-    ExitIndex = len(CharacterClass.Classes)
-    while True: 
-        text = Engine.Text("")
-        for index , _ in enumerate(CharacterClass.Classes):
-            if TargetCharacter == index: text.append(f"> {CharacterClass.Classes[index].Name}\n" ,style="u %s" %HIGH_LIGHT_COLOR)
-            else: text.append(f"{CharacterClass.Classes[index].Name}\n")
-        if TargetCharacter == ExitIndex: text.append("> Go to menu" ,style="u %s" %HIGH_LIGHT_COLOR)
-        else: text.append("Go to menu")
-
-        Engine.layout['Main'].update(Engine.Panel(text,title="Character Select",style="%s" %PANEL_COLOR))
-        match Game.get_input():
-            case 'w' :
-                TargetCharacter -= 1
-                if TargetCharacter < 0: TargetCharacter = ExitIndex
-            case 's' :
-                TargetCharacter += 1
-                if TargetCharacter > ExitIndex: TargetCharacter = 0
-            case '\r' | ' ':
-                if TargetCharacter == ExitIndex: return False
-
-                if TargetCharacter in range(len(CharacterClass.Classes)+1):
-                    player.SetAllStats(CharacterClass.Classes[TargetCharacter].Stats , CharacterClass.Classes[TargetCharacter].Name)
-                    Engine.layout['Side'].update(Engine.Panel(Engine.Text(f"You've choosen {CharacterClass.Classes[TargetCharacter].Name}",style="green"),style="%s" %PANEL_COLOR))
-                    time.sleep(1)
-                    return True
-                else:
-                    Engine.layout['Side'].update(Engine.Panel(Engine.Text("Character doesn't exist!",style="red"),style="%s" %PANEL_COLOR))
-
-def RollNextEnemy() -> Monster:
-    value = random.randint(0,len(Monster.MonsterBase)-1)
-    return Monster.MonsterBase[value]
-
-def NextTurn():
-    MainText = Engine.Text("You took a next turn")
-    Engine.layout['Main'].update(Engine.Panel(MainText,title=f"Health {player.GetHealth()}/{player.GetMaxHealth()}",style="%s" %PANEL_COLOR))
-    rng = random.randint(0,100)
-    SideText = Engine.Text("",justify="center")
-    if rng in range(21,45):
-        SideText.append('You found a place to rest\n')
-        ProcentHealthRested = random.randint(5,25)/100
-        HpRested = round(player.GetMaxHealth() * ProcentHealthRested)
-        player.SetHealth(player.GetHealth()+HpRested)
-        SideText.append("And recoverd %s Health\n" % HpRested)
-        SideText.append("Current Health %s/%s" %(round(player.GetHealth()),player.GetMaxHealth()))
-        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-    elif rng in range(46,80):
-        RandomEnemy = RollNextEnemy()
-        TargetOption = 0
-        while True:
-            SideText = Engine.Text("")
-            SideText.append('On your way you encounter a %s \n' %RandomEnemy.name)
-            SideText.append('Do you want to fight?\n')
-            AnswerText = Engine.Text("")
-            if TargetOption == 0: AnswerText.append('Yes/No').stylize("u honeydew2",0,3)
-            else: AnswerText.append('Yes/No').stylize("u honeydew2",4)
-            
-            Engine.layout['Side'].update(Engine.Panel(Engine.Text.assemble(SideText,AnswerText,justify="center"),style="%s" %PANEL_COLOR))
-            match Game.get_input():
-                case 'w' :
-                    TargetOption -= 1
-                    if TargetOption < 0: TargetOption = 1
-                case 's' :
-                    TargetOption += 1
-                    if TargetOption > 1: TargetOption = 0
-                case "\r" | " ":
-                    if TargetOption == 0:
-                        Fight(RandomEnemy)
-                        return
-                    EscapeChance = random.randint(0,2)
-                    if EscapeChance == 0:
-                        Engine.layout['Side'].update(Engine.Panel(Engine.Text("You failed to escape from a %s" %RandomEnemy.name),style="%s" %PANEL_COLOR))
-                        Fight(RandomEnemy)
-                        return
-                    Engine.layout['Side'].update(Engine.Panel(Engine.Text("You've escaped from a %s" %RandomEnemy.name),style="%s" %PANEL_COLOR))
-                    return
-    elif rng in range(81,95):
-        player.Potions['Health Potion'] += 1
-        SideText.append('Wow! you found a Healing potion\nYou now have %s' % player.Potions['Health Potion'])
-        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-    elif rng in range(96,100):
-        rngItem = random.randint(0,len(Item.ItemBase)-1)
-        founditem = Item.ItemBase[rngItem]
-        player.Inventory.append(founditem)
-        SideText.append('You found an %s lying on the ground' % founditem.name)
-        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-    else: Engine.layout['Side'].update(Engine.Panel(Engine.Text("Nothing excites happens on your journey",justify="center"),style="%s" %PANEL_COLOR))
-    
-def Fight(Monster:Monster):
-    Engine.layout['Side'].update(Engine.Panel(GameNamePrint(),style="%s" %PANEL_COLOR))
-    Options = ["Attack","Inventory","Equipment","Run Away"]
-    TargetOption = 0
-    ExitIndex = len(Options)
-    enemy = Monster
-    enemy.SetMaxHealth()
-    while True:
-        if player.GetHealth() <= 0:
-            Engine.layout['Main'].update(Engine.Panel(GameNamePrint(),title="Game",style="%s" %PANEL_COLOR))
-            Engine.layout['Side'].update(Engine.Panel(Engine.Text("You died!",style="red"),style="%s" %PANEL_COLOR))
-            Game.wait_for_input()
-            return 0
-        MainText = Engine.Text("")
-        SideText = Engine.Text("")
-
-        for index , Option in enumerate(Options):
-            if TargetOption == index: MainText.append("> %s \n" %Option ,style="u %s" %HIGH_LIGHT_COLOR)
-            else: MainText.append("%s \n" %Option)
-
-        SideText.append("You're now fighting %s \n"%enemy.name)
-        SideText.append("%s/%s Health \n" %(enemy.Stats['Health'],enemy.Stats['MaxHealth']))
-        SideText.append("%s Attack \n" %enemy.Stats['Attack'])
-        SideText.append("%s Defence \n\n" %enemy.Stats['Defence'])
-        SideText.append(player.PrintStats())
-        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-        Engine.layout['Main'].update(Engine.Panel(MainText,title="Fight",style="%s" %PANEL_COLOR))
-
-        match Game.get_input():
-            case 'w' :
-                TargetOption -= 1
-                if TargetOption < 0: TargetOption = ExitIndex
-            case 's' :
-                TargetOption += 1
-                if TargetOption > ExitIndex: TargetOption = 0
-            case '\r' | ' ':
-                match TargetOption:
-                    case 0:
-                        SideText = Engine.Text("")
-                        SideText.append("You've dealt %s damage \n" %CalculateDamage(player.GetAttack() ,enemy.Stats['Defence']))
-                        enemy.SetHealth(enemy.GetHealth() - CalculateDamage(player.GetAttack() ,enemy.Stats['Defence']))
-                        SideText.append('%s Health %s/%s \n' %(enemy.name ,enemy.GetHealth() ,enemy.Stats['MaxHealth']))
-                        if enemy.Stats['Health'] <= 0:
-                            SideText.append("You've killed %s \n" %enemy.name)
-                            dropped = enemy.LootTable.Roll()
-                            if dropped != None:   
-                                SideText.append("It dropped a %s" %dropped)
-                            else:
-                                SideText.append("It didn't drop anything")
-                            Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-                            return 0
-                        player.SetHealth(player.GetHealth()-CalculateDamage(enemy.Stats['Attack'],player.GetDefence()))
-                        SideText.append("\n%s dealt %s damage \n" %(enemy.name,CalculateDamage(enemy.Stats['Attack'],player.GetDefence())))
-                        SideText.append("Player's Health %s/%s \n" %(player.Stats['Health'],player.Stats['MaxHealth']))
-                        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-                        Game.wait_for_input()
-                    case 1:
-                        ShowInventory()
-                    case 2:
-                        ShowEquipment()
-                    case ExitIndex:
-                        x = random.randint(0,1)
-                        if x == 0:
-                            Engine.layout['Side'].update(Engine.Panel(Engine.Text("You've escaped from %s" %enemy.name),style="%s" %PANEL_COLOR))
-                            return 0
-                        else:
-                            SideText = Engine.Text("")
-                            SideText.append('You failled to escape')
-                            player.SetHealth(player.GetHealth()-CalculateDamage(enemy.Stats['Attack'],player.GetDefence()))
-                            SideText.append("\n%s dealt %s damage " %(enemy.name,CalculateDamage(enemy.Stats['Attack'],player.GetDefence())))
-                            SideText.append("Player's Health %s/%s " %(player.Stats['Health'],player.Stats['MaxHealth']))
-                            Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
-                            Game.wait_for_input()
-
-def CalculateDamage(Damage:int,Defence:int):
-    reduction = 1+Defence/100
-    result = Damage/reduction
-    return int(result)
-
-def Play():
-    Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
-    TargetOption = 0
-    Options = ['Next Turn','Shop (WIP)','Inventory','Equipment','Go to menu']
-    ExitIndex = len(Options)-1
-    while True:
-        if player.GetHealth() <= 0: 
-            player.Die()
-            return 0
-        MainText = Engine.Text("")
-        for index , name in enumerate(Options):
-            if index == TargetOption: MainText.append("> %s \n" %name ,style="u %s" %HIGH_LIGHT_COLOR)
-            else: MainText.append("%s \n" %name)
-
-        Engine.layout['Main'].update(Engine.Panel(MainText,title="Game",style="%s" %PANEL_COLOR))
-        match Game.get_input():
-            case 'w' :
-                TargetOption -= 1
-                if TargetOption < 0: TargetOption = ExitIndex
-            case 's' :
-                TargetOption += 1
-                if TargetOption > ExitIndex: TargetOption = 0
-            case '\r' | ' ':
-                match TargetOption:
-                    case 0:
-                        AutoSave()
-                        NextTurn()
-                        Game.wait_for_input()
-                        Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
-                    case 1:
-                        pass
-                    case 2:
-                        ShowInventory()
-                        Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
-                    case 3:
-                        ShowEquipment()
-                        Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
-                    case ExitIndex:
-                        answer = 0
-                        while True:
-                            text = Engine.Text("Do you want to save before quiting?\n")
-                            answerText = Engine.Text("")
-                            if answer == 0: answerText.append("Yes/No").stylize("u honeydew2",0,3)
-                            else: answerText.append("Yes/No").stylize("u honeydew2",4)
-                            Engine.layout['Side'].update(Engine.Panel(Engine.Text.assemble(text,answerText,justify='center'),style="%s" %PANEL_COLOR))
-                            match Game.get_input():
-                                case 'w' :
-                                    answer -= 1
-                                    if answer < 0: answer = 1
-                                case 's' :
-                                    answer += 1
-                                    if answer > 1: answer = 0
-                                case '\r' | ' ':
-                                    match answer:
-                                        case 0:
-                                            Engine.layout['Side'].update(Engine.Panel(Engine.Text("Type a save file name",justify='center'),style="%s" %PANEL_COLOR))
-                                            SaveFileName = Engine.Text("")
-                                            while True:
-                                                userInput = Engine.msvcrt.getwch()
-                                                if userInput == '\r': break
-                                                SaveFileName = Engine.Text(Game.TextBoxInput(SaveFileName,userInput))
-                                                HelpText = Engine.Text("Type a save file name\n")
-                                                Engine.layout['Side'].update(Engine.Panel(Engine.Text.assemble(HelpText,SaveFileName,justify="center"),style="%s" %PANEL_COLOR))
-                                            if str(SaveFileName) == "":
-                                                if player.CurrentSaveFileName == "": 
-                                                    player.CurrentSaveFileName = ('AutoSave-%s-%s' %(player.Class,date.today()))
-                                                AutoSave()
-                                            else: Save(str(SaveFileName))
-                                            player.Die()
-                                            return 0
-                                        case 1:
-                                            player.Die()
-                                            return 0
-
-def GameNamePrint(Style:str = "spring_green2"):
-    text = Engine.Text("",justify="center")
-    text.append('____ ___  ____    ____ ____ _  _ ____ \n',style=f'{Style}')
-    text.append('|__/ |__] | __    | __ |__| |\/| |___ \n',style=f'{Style}')
-    text.append('|  \ |    |__]    |__] |  | |  | |___ \n',style=f'{Style}')
-    return text
-
-def Menu():
-    options = ['Play','Load','Development Info','Exit']
-    TargetOption = 0
-    ExitIndex = len(options)-1
-    while True:
-        MainText = Engine.Text("")
-        Engine.layout['Side'].update(Engine.Panel(GameNamePrint(),style="%s" %PANEL_COLOR))
-        for index, name in enumerate(options):
-            if TargetOption == index: MainText.append("> %s \n" %name ,style="u %s" %HIGH_LIGHT_COLOR)
-            else: MainText.append("%s \n" %name)
-
-        Engine.layout['Main'].update(Engine.Panel(MainText,title="Menu" ,style="%s" %PANEL_COLOR))
-        match Game.get_input():
-            case 'w' | "\r'P'":
-                TargetOption -= 1
-                if TargetOption < 0: TargetOption = ExitIndex
-            case 's' :
-                TargetOption += 1
-                if TargetOption > ExitIndex: TargetOption = 0
-            case '\r' | ' ':
-                match TargetOption:
-                    case 0:
-                        if ChooseCharacter():
-                            player.CurrentSaveFileName = ('AutoSave-%s-%s' %(player.Class,date.today()))
-                            Play()
-                    case 1:
-                        if SavesMenu() == True:
-                            Play()
-                    case 2:
-                        Engine.layout['Side'].update(Engine.Panel(DeveloperInfoMenu(),style="%s" %PANEL_COLOR))
-                        Game.wait_for_input()
-                    case ExitIndex:
-                        sys.exit()
-    return
 
 def SavesMenu():
     saves = []
@@ -845,7 +259,6 @@ def Save(SaveName:str):
 
 def Load(SaveName:str):
     player.CurrentSaveFileName = SaveName.split(".json")[0]
-    Game.Clear()
     try:
         with open(SaveFilePath+"/"+SaveName+".json", "r") as read_file:
             data = json.load(read_file)
@@ -872,25 +285,368 @@ def Load(SaveName:str):
                 player.Inventory.append(y)
     return 1
 
-def Debug():
+def Shop():
     pass
 
+def ShowInventory():    # TODO ADD SHOWING ITEM'S STATS ON A MAIN SCREEN WHILE IT IS TARGETED
+    TargetOption = 0
+    TargetItem = None
+    while True:
+        ExitIndex = len(player.Potions) + len(player.Inventory)
+        SideText = Engine.Text("")
+        SideText.append("Potions: \n")
+        for index, name in enumerate(player.Potions):
+            if index == TargetOption: 
+                SideText.append("> %s - %s\n" % (name,player.Potions[name]) ,style="u %s" %HIGH_LIGHT_COLOR)
+                TargetItem = name
+            else: SideText.append("%s - %s\n" % (name,player.Potions[name]))
+        SideText.append("Items: \n")
+        if len(player.Inventory) > 0:
+            for index, Item in enumerate(player.Inventory):
+                if index+len(player.Potions) == TargetOption: 
+                    SideText.append("> %s \n" % Item.name ,style="u %s" %HIGH_LIGHT_COLOR)
+                    TargetItem = Item.name
+                else: SideText.append("%s \n" % Item.name)
+            
+        if TargetOption == ExitIndex: SideText.append("> Go back \n" ,style="u %s" %HIGH_LIGHT_COLOR)
+        else: SideText.append("Go back \n")
+
+        Engine.layout['Main'].update(Engine.Panel(player.PrintStats(),title="Player",style="%s" %PANEL_COLOR))
+        Engine.layout['Side'].update(Engine.Panel(SideText,title="Inventory",style="%s" %PANEL_COLOR))
+        match Game.get_input():
+            case 'w' :
+                TargetOption -= 1
+                if TargetOption < 0: TargetOption = ExitIndex
+            case 's' :
+                TargetOption += 1
+                if TargetOption > ExitIndex: TargetOption = 0
+            case '\r' | ' ':
+                if TargetOption == ExitIndex: return
+                if TargetItem != None:
+                    player.UsePotion(TargetItem)
+                    player.UseItem(TargetItem)
+
+def ShowEquipment():
+    TargetOption = 0
+    TargetItem = None
+    ExitIndex = len(player.Equiped)
+    while True:
+        EquipmentText = Engine.Text("")
+        for index , name in enumerate(player.Equiped):
+            Item = player.Equiped[name]
+            if TargetOption == index:
+                EquipmentText.append("> %s - %s \n" %(name,Item) ,style="u %s" %HIGH_LIGHT_COLOR)
+                TargetItem = Item
+            else: EquipmentText.append("%s - %s \n" %(name,Item))
+        if TargetOption == ExitIndex: EquipmentText.append("> Go back" ,style="u %s" %HIGH_LIGHT_COLOR)
+        else: EquipmentText.append("Go back")
+        Engine.layout['Main'].update(Engine.Panel(player.PrintStats(),title="Player",style="%s" %PANEL_COLOR))
+        Engine.layout['Side'].update(Engine.Panel(EquipmentText,title="Equipement",style="%s" %PANEL_COLOR))
+        match Game.get_input():
+            case 'w' :
+                TargetOption -= 1
+                if TargetOption < 0: TargetOption = ExitIndex
+            case 's' :
+                TargetOption += 1
+                if TargetOption > ExitIndex: TargetOption = 0
+            case '\r' | ' ':
+                if TargetOption == ExitIndex: return
+                if TargetItem != None:
+                    Engine.layout['Side'].update(Engine.Panel(player.RemoveEquipment(TargetItem),title="Equipement",style="%s" %PANEL_COLOR))
+                    time.sleep(1)
+
+def ChooseCharacter():
+    Engine.layout['Side'].update(Engine.Panel(Engine.Text("Choose your character!"),style="%s" %PANEL_COLOR))
+    TargetCharacter = 0
+    ExitIndex = len(CharacterClass.Classes)
+    while True: 
+        text = Engine.Text("")
+        for index , _ in enumerate(CharacterClass.Classes):
+            if TargetCharacter == index: text.append(f"> {CharacterClass.Classes[index].Name}\n" ,style="u %s" %HIGH_LIGHT_COLOR)
+            else: text.append(f"{CharacterClass.Classes[index].Name}\n")
+        if TargetCharacter == ExitIndex: text.append("> Go to menu" ,style="u %s" %HIGH_LIGHT_COLOR)
+        else: text.append("Go to menu")
+
+        Engine.layout['Main'].update(Engine.Panel(text,title="Character Select",style="%s" %PANEL_COLOR))
+        match Game.get_input():
+            case 'w' :
+                TargetCharacter -= 1
+                if TargetCharacter < 0: TargetCharacter = ExitIndex
+            case 's' :
+                TargetCharacter += 1
+                if TargetCharacter > ExitIndex: TargetCharacter = 0
+            case '\r' | ' ':
+                if TargetCharacter == ExitIndex: return False
+
+                if TargetCharacter in range(len(CharacterClass.Classes)+1):
+                    player.SetAllStats(CharacterClass.Classes[TargetCharacter].Stats , CharacterClass.Classes[TargetCharacter].Name)
+                    Engine.layout['Side'].update(Engine.Panel(Engine.Text(f"You've choosen {CharacterClass.Classes[TargetCharacter].Name}",style="green"),style="%s" %PANEL_COLOR))
+                    time.sleep(1)
+                    return True
+                else:
+                    Engine.layout['Side'].update(Engine.Panel(Engine.Text("Character doesn't exist!",style="red"),style="%s" %PANEL_COLOR))
+
+def CalculateDamage(Damage:int,Defence:int):
+    reduction = 1+Defence/100
+    result = Damage/reduction
+    return int(result)
+
+def Fight(Monster):
+    Engine.layout['Side'].update(Engine.Panel(GameNamePrint(),style="%s" %PANEL_COLOR))
+    Options = ["Attack","Inventory","Equipment","Run Away"]
+    TargetOption = 0
+    ExitIndex = len(Options)
+    enemy = Monster
+    enemy.SetMaxHealth()
+    while True:
+        if player.GetHealth() <= 0:
+            Engine.layout['Main'].update(Engine.Panel(GameNamePrint(),title="Game",style="%s" %PANEL_COLOR))
+            Engine.layout['Side'].update(Engine.Panel(Engine.Text("You died!",style="red"),style="%s" %PANEL_COLOR))
+            Game.wait_for_input()
+            return 0
+        MainText = Engine.Text("")
+        SideText = Engine.Text("")
+
+        for index , Option in enumerate(Options):
+            if TargetOption == index: MainText.append("> %s \n" %Option ,style="u %s" %HIGH_LIGHT_COLOR)
+            else: MainText.append("%s \n" %Option)
+
+        SideText.append("You're now fighting %s \n"%enemy.name)
+        SideText.append("%s/%s Health \n" %(enemy.Stats['Health'],enemy.Stats['MaxHealth']))
+        SideText.append("%s Attack \n" %enemy.Stats['Attack'])
+        SideText.append("%s Defence \n\n" %enemy.Stats['Defence'])
+        SideText.append(player.PrintStats())
+        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+        Engine.layout['Main'].update(Engine.Panel(MainText,title="Fight",style="%s" %PANEL_COLOR))
+
+        match Game.get_input():
+            case 'w' :
+                TargetOption -= 1
+                if TargetOption < 0: TargetOption = ExitIndex
+            case 's' :
+                TargetOption += 1
+                if TargetOption > ExitIndex: TargetOption = 0
+            case '\r' | ' ':
+                match TargetOption:
+                    case 0:
+                        SideText = Engine.Text("")
+                        SideText.append("You've dealt %s damage \n" %CalculateDamage(player.GetAttack() ,enemy.Stats['Defence']))
+                        enemy.SetHealth(enemy.GetHealth() - CalculateDamage(player.GetAttack() ,enemy.Stats['Defence']))
+                        SideText.append('%s Health %s/%s \n' %(enemy.name ,enemy.GetHealth() ,enemy.Stats['MaxHealth']))
+                        if enemy.Stats['Health'] <= 0:
+                            SideText.append("You've killed %s \n" %enemy.name)
+                            dropped = enemy.LootTable.RollForItem(player)
+                            if dropped != None:   
+                                SideText.append("It dropped a %s" %dropped)
+                            else:
+                                SideText.append("It didn't drop anything")
+                            Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+                            return 0
+                        player.SetHealth(player.GetHealth()-CalculateDamage(enemy.Stats['Attack'],player.GetDefence()))
+                        SideText.append("\n%s dealt %s damage \n" %(enemy.name,CalculateDamage(enemy.Stats['Attack'],player.GetDefence())))
+                        SideText.append("Player's Health %s/%s \n" %(player.Stats['Health'],player.Stats['MaxHealth']))
+                        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+                        Game.wait_for_input()
+                    case 1:
+                        ShowInventory()
+                    case 2:
+                        ShowEquipment()
+                    case ExitIndex:
+                        x = random.randint(0,1)
+                        if x == 0:
+                            Engine.layout['Side'].update(Engine.Panel(Engine.Text("You've escaped from %s" %enemy.name),style="%s" %PANEL_COLOR))
+                            return 0
+                        else:
+                            SideText = Engine.Text("")
+                            SideText.append('You failled to escape')
+                            player.SetHealth(player.GetHealth()-CalculateDamage(enemy.Stats['Attack'],player.GetDefence()))
+                            SideText.append("\n%s dealt %s damage " %(enemy.name,CalculateDamage(enemy.Stats['Attack'],player.GetDefence())))
+                            SideText.append("Player's Health %s/%s " %(player.Stats['Health'],player.Stats['MaxHealth']))
+                            Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+                            Game.wait_for_input()
+
+def NextTurn():
+    MainText = Engine.Text("You took a next turn")
+    Engine.layout['Main'].update(Engine.Panel(MainText,title=f"Health {player.GetHealth()}/{player.GetMaxHealth()}",style="%s" %PANEL_COLOR))
+    rng = random.randint(0,100)
+    SideText = Engine.Text("",justify="center")
+    if rng in range(21,45):
+        SideText.append('You found a place to rest\n')
+        ProcentHealthRested = random.randint(5,25)/100
+        HpRested = round(player.GetMaxHealth() * ProcentHealthRested)
+        player.SetHealth(player.GetHealth()+HpRested)
+        SideText.append("And recoverd %s Health\n" % HpRested)
+        SideText.append("Current Health %s/%s" %(round(player.GetHealth()),player.GetMaxHealth()))
+        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+    elif rng in range(46,80):
+        RandomEnemy = Monster.RollNextEnemy()
+        TargetOption = 0
+        while True:
+            SideText = Engine.Text("")
+            SideText.append('On your way you encounter a %s \n' %RandomEnemy.name)
+            SideText.append('Do you want to fight?\n')
+            AnswerText = Engine.Text("")
+            if TargetOption == 0: AnswerText.append('Yes/No').stylize("u %s" %HIGH_LIGHT_COLOR ,0,3)
+            else: AnswerText.append('Yes/No').stylize("u %s" %HIGH_LIGHT_COLOR ,4)
+            
+            Engine.layout['Side'].update(Engine.Panel(Engine.Text.assemble(SideText,AnswerText,justify="center"),style="%s" %PANEL_COLOR))
+            match Game.get_input():
+                case 'w' :
+                    TargetOption -= 1
+                    if TargetOption < 0: TargetOption = 1
+                case 's' :
+                    TargetOption += 1
+                    if TargetOption > 1: TargetOption = 0
+                case "\r" | " ":
+                    if TargetOption == 0:
+                        Fight(RandomEnemy)
+                        return
+                    EscapeChance = random.randint(0,2)
+                    if EscapeChance == 0:
+                        Engine.layout['Side'].update(Engine.Panel(Engine.Text("You failed to escape from a %s" %RandomEnemy.name),style="%s" %PANEL_COLOR))
+                        Fight(RandomEnemy)
+                        return
+                    Engine.layout['Side'].update(Engine.Panel(Engine.Text("You've escaped from a %s" %RandomEnemy.name),style="%s" %PANEL_COLOR))
+                    return
+    elif rng in range(81,95):
+        player.Potions['Health Potion'] += 1
+        SideText.append('Wow! you found a Healing potion\nYou now have %s' % player.Potions['Health Potion'])
+        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+    elif rng in range(96,100):
+        rngItem = random.randint(0,len(Item.ItemBase)-1)
+        founditem = Item.ItemBase[rngItem]
+        player.Inventory.append(founditem)
+        SideText.append('You found an %s lying on the ground' % founditem.name)
+        Engine.layout['Side'].update(Engine.Panel(SideText,style="%s" %PANEL_COLOR))
+    else: Engine.layout['Side'].update(Engine.Panel(Engine.Text("Nothing excites happens on your journey",justify="center"),style="%s" %PANEL_COLOR))
+
+def Play():
+    Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
+    TargetOption = 0
+    Options = ['Next Turn','Shop (WIP)','Inventory','Equipment','Go to menu']
+    ExitIndex = len(Options)-1
+    while True:
+        if player.GetHealth() <= 0: 
+            player.Die()
+            return 0
+        MainText = Engine.Text("")
+        for index , name in enumerate(Options):
+            if index == TargetOption: MainText.append("> %s \n" %name ,style="u %s" %HIGH_LIGHT_COLOR)
+            else: MainText.append("%s \n" %name)
+
+        Engine.layout['Main'].update(Engine.Panel(MainText,title="Game",style="%s" %PANEL_COLOR))
+        match Game.get_input():
+            case 'w' :
+                TargetOption -= 1
+                if TargetOption < 0: TargetOption = ExitIndex
+            case 's' :
+                TargetOption += 1
+                if TargetOption > ExitIndex: TargetOption = 0
+            case '\r' | ' ':
+                match TargetOption:
+                    case 0:
+                        AutoSave()
+                        NextTurn()
+                        Game.wait_for_input()
+                        Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
+                    case 1:
+                        pass
+                    case 2:
+                        ShowInventory()
+                        Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
+                    case 3:
+                        ShowEquipment()
+                        Engine.layout['Side'].update(Engine.Panel(player.PrintStats(),style="%s" %PANEL_COLOR))
+                    case ExitIndex:
+                        answer = 0
+                        while True:
+                            text = Engine.Text("Do you want to save before quiting?\n")
+                            answerText = Engine.Text("")
+                            if answer == 0: answerText.append("Yes/No").stylize("u %s" %HIGH_LIGHT_COLOR ,0,3)
+                            else: answerText.append("Yes/No").stylize("u %s" %HIGH_LIGHT_COLOR ,4)
+                            Engine.layout['Side'].update(Engine.Panel(Engine.Text.assemble(text,answerText,justify='center'),style="%s" %PANEL_COLOR))
+                            match Game.get_input():
+                                case 'w' :
+                                    answer -= 1
+                                    if answer < 0: answer = 1
+                                case 's' :
+                                    answer += 1
+                                    if answer > 1: answer = 0
+                                case '\r' | ' ':
+                                    match answer:
+                                        case 0:
+                                            Engine.layout['Side'].update(Engine.Panel(Engine.Text("Type a save file name",justify='center'),style="%s" %PANEL_COLOR))
+                                            SaveFileName = Engine.Text("")
+                                            while True:
+                                                userInput = Engine.msvcrt.getwch()
+                                                if userInput == '\r': break
+                                                SaveFileName = Engine.Text(Game.TextBoxInput(SaveFileName,userInput))
+                                                HelpText = Engine.Text("Type a save file name\n")
+                                                Engine.layout['Side'].update(Engine.Panel(Engine.Text.assemble(HelpText,SaveFileName,justify="center"),style="%s" %PANEL_COLOR))
+                                            if str(SaveFileName) == "":
+                                                if player.CurrentSaveFileName == "": 
+                                                    player.CurrentSaveFileName = ('AutoSave-%s-%s' %(player.Class,date.today()))
+                                                AutoSave()
+                                            else: Save(str(SaveFileName))
+                                            player.Die()
+                                            return 0
+                                        case 1:
+                                            player.Die()
+                                            return 0
+
+def GameNamePrint(Style:str = "spring_green2"):
+    text = Engine.Text("",justify="center")
+    text.append('____ ___  ____    ____ ____ _  _ ____ \n',style=f'{Style}')
+    text.append('|__/ |__] | __    | __ |__| |\/| |___ \n',style=f'{Style}')
+    text.append('|  \ |    |__]    |__] |  | |  | |___ \n',style=f'{Style}')
+    return text
+
+def Menu():
+    options = ['Play','Load','Development Info','Exit']
+    TargetOption = 0
+    ExitIndex = len(options)-1
+    while True:
+        MainText = Engine.Text("")
+        Engine.layout['Side'].update(Engine.Panel(GameNamePrint(),style="%s" %PANEL_COLOR))
+        for index, name in enumerate(options):
+            if TargetOption == index: MainText.append("> %s \n" %name ,style="u %s" %HIGH_LIGHT_COLOR)
+            else: MainText.append("%s \n" %name)
+
+        Engine.layout['Main'].update(Engine.Panel(MainText,title="Menu" ,style="%s" %PANEL_COLOR))
+        match Game.get_input():
+            case 'w' | "\r'P'":
+                TargetOption -= 1
+                if TargetOption < 0: TargetOption = ExitIndex
+            case 's' :
+                TargetOption += 1
+                if TargetOption > ExitIndex: TargetOption = 0
+            case '\r' | ' ':
+                match TargetOption:
+                    case 0:
+                        if ChooseCharacter():
+                            player.CurrentSaveFileName = ('AutoSave-%s-%s' %(player.Class,date.today()))
+                            Play()
+                    case 1:
+                        if SavesMenu() == True:
+                            Play()
+                    case 2:
+                        Engine.layout['Side'].update(Engine.Panel(DeveloperInfoMenu(),style="%s" %PANEL_COLOR))
+                        Game.wait_for_input()
+                    case ExitIndex:
+                        sys.exit()
+    return
 
 def DeveloperInfoMenu():
     text = Engine.Text("")
     text.append("Development info\n\n",style="cyan")
     text.append("Game Roadmap:\n",style="deep_sky_blue4")
-    text.append("1.Add shops and currency\n")
-    text.append("2.Balance current gameplay\n")
-    text.append("3.Add more Items and Monsters\n")
-    text.append("4.Add Leveling system and exp gathering from fights\n")
-    text.append("5.Expand loot tables and refactor it's code\n\n")
+    text.append("1.Add shops and currency\n2.Balance current gameplay\n3.Add more Items and Monsters\n",
+                "4.Add Leveling system and exp gathering from fights\n5.Expand loot tables and refactor it's code\n\n")
     text.append("Change Log:\n",style="spring_green4")
-    text.append("Changed the game to work with the new GUI\nfixed some crashes\nfor now removed selling items and gold\n\n")
-    text.append("Currently adding shops and merchants\nMerchants you will be able to encounter in the wilderness. They can sell you some really good items but they are really rare\n\n")
+    text.append("Changed the game to work with the new GUI\nfixed some crashes\nfor now removed selling items and gold\n\n",
+    "Currently adding shops and merchants\nMerchants you will be able to encounter in the wilderness. They can sell you some really good items but they are really rare\n\n")
     text.append('Press any button to continue',style="blink bright_black")
     return text
-
 
 
 if __name__ == '__main__':
